requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
waitress==2.1.2
pylint==2.17.5
python-dotenv==1.0.0  